fix: resolve 15 mypy strict errors across 7 files
instrument.py   add return type Callable[[type], type], annotate
                _orig: Any and _name: str default params
meta_harness.py str() cast on dict.get() to fix no-any-return
scan/metrics.py float() cast on numpy floating[Any] accumulation
collect/
  collector.py  __exit__ -> None (was bool); remove return False
  study_client  _orig: Any, _name: str on inner wrapper defaults
optimize/
  roadmap.py    add Any import; dict[str, Any] on class var and
                _default_phases return type
cli.py          TYPE_CHECKING import for EpisodeRecord; annotate
                _load_jsonl -> list[dict[str, Any]] with explicit
                intermediate variable; fix raw_records / failed_raw
                annotation; type helper function signatures

+++ src/hachillesworld/scan/metrics.py
@@ -279,5 +279,5 @@ class MetricsCalculator:
                 continue
             avg_conf = np.mean([c for c, _ in in_bin])
             avg_acc = np.mean([a for _, a in in_bin])
-            ece += (len(in_bin) / n) * abs(avg_conf - avg_acc)
+            ece += float((len(in_bin) / n) * abs(avg_conf - avg_acc))
         return float(ece)
